+++ pyslvs_ui/entities/__init__.py
@@ -2,7 +2,6 @@
 
 """'entities' module contains the option dialogs."""
 
-__all__ = ['EditLinkDialog', 'EditPointDialog']
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"