+++ pyslvs_ui/entities/edit_link.py
@@ -11,15 +11,11 @@ from re import match
 from typing import List, Union
 from qtpy.QtCore import Slot, Qt
 from qtpy.QtWidgets import (
-    QDialog,
-    QListWidgetItem,
-    QDialogButtonBox,
-    QColorDialog,
-    QWidget,
+    QDialog, QListWidgetItem, QDialogButtonBox, QColorDialog, QWidget,
 )
 from qtpy.QtGui import QIcon, QPixmap
-from pyslvs import VPoint, VLink
-from pyslvs_ui.graphics import color_names, color_qt, color_icon
+from pyslvs import VPoint, VLink, color_names
+from pyslvs_ui.graphics import color_qt, color_icon
 from .utility import set_custom_color, add_custom_color
 from .edit_link_ui import Ui_Dialog