+++ pyslvs_ui/synthesis/__init__.py
@@ -2,13 +2,6 @@
 
 """'synthesis' module contains synthesis functional interfaces."""
 
-__all__ = [
-    'StructureSynthesis',
-    'Collections',
-    'CollectionsDialog',
-    'ConfigureWidget',
-    'DimensionalSynthesis'
-]
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"