+++ pyslvs_ui/synthesis/dimensional_synthesis/dialogs/__init__.py
@@ -2,13 +2,6 @@
 
 """'dialogs' module contains contains the dialog of this tab."""
 
-__all__ = [
-    'AlgorithmOptionDialog',
-    'EditPathDialog',
-    'ProgressDialog',
-    'PreviewDialog',
-    'ChartDialog'
-]
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2020"
 __license__ = "AGPL"